Time: 270 ms (60.85%), Space: 85.8 MB (40.69%) - LeetHub

+++ 853-car-fleet/853-car-fleet.cpp
@@ -1,14 +1,14 @@
 class Solution {
 public:
     int carFleet(int target, vector<int>& position, vector<int>& speed) {
-        priority_queue<pair<int,float>> pq;
+        priority_queue<pair<int,double>> pq;
         for(int i=0;i<position.size();i++){
-            float time = (target-position[i])/(float)speed[i];
+            double time = (target-position[i])/(double)speed[i];
             pq.push({position[i], time});
         }
         int count = 0;
         while(!pq.empty()){
-            float currTime = pq.top().second;
+            double currTime = pq.top().second;
             while(!pq.empty() && pq.top().second<=currTime){
                 pq.pop();
             }
